Changed PID Values
I fixed the P-value for ctwg() and cdwg() and added in an I-value for the wallaby turn_with_gyro(). I still haven't added d-values for either twg() or ctwg() and haven't added an I-value for ctwg(). I added anti-windup code to both functions as well fixing the d controller (It now update previous_error).